created needed os context annotation and added getManager to addon interface
