+++ OSA/addon/android/fermat-osa-addon-android-database-system-bitdubai/src/main/java/com/bitdubai/fermat_osa_addon/layer/android/database_system/developer/bitdubai/version_1/PlatformDatabaseSystemAndroidAddonRoot.java
@@ -1,25 +1,14 @@
 package com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1;
 
 import android.content.Context;
-import android.util.Base64;
 
 import com.bitdubai.fermat_api.CantStartPluginException;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.abstract_classes.AbstractAddon;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededOsContext;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.AddonVersionReference;
-import com.bitdubai.fermat_api.layer.all_definition.enums.ServiceStatus;
 import com.bitdubai.fermat_api.layer.all_definition.util.Version;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.Database;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.PlatformDatabaseSystem;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantCreateDatabaseException;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantOpenDatabaseException;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.DatabaseNotFoundException;
-import com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1.exceptions.CantHashDatabaseNameException;
-import com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1.structure.AndroidDatabase;
-
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.Charset;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
+import com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1.structure.AndroidPlatformDatabaseSystem;
 
 /**
  * This addon handles a layer of database representation.
@@ -29,130 +18,38 @@ import java.security.NoSuchAlgorithmException;
  * Created by lnacosta (laion.cj91@gmail.com) on 26/10/2015.
  */
 
-public class PlatformDatabaseSystemAndroidAddonRoot extends AbstractAddon implements PlatformDatabaseSystem {
+public class PlatformDatabaseSystemAndroidAddonRoot extends AbstractAddon {
 
+    @NeededOsContext
     private Context context;
 
+    private FermatManager platformDatabaseSystemManager;
+
+    /**
+     * Constructor without parameters.
+     */
     public PlatformDatabaseSystemAndroidAddonRoot() {
-        super(
-                new AddonVersionReference(new Version()),
-                true
-        );
+        super(new AddonVersionReference(new Version()));
     }
 
     @Override
-    public void start() throws CantStartPluginException {
-
-        if (this.getOsContext() != null && this.getOsContext() instanceof Context) {
-            context = (Context) this.getOsContext();
-            this.serviceStatus = ServiceStatus.STARTED;
-        } else {
-            throw new CantStartPluginException(
-                    "osContext: "+this.getOsContext(),
-                    "Context is not instance of Android Context or is null."
-            );
-        }
-    }
-
-    @Override
-    public final Database openDatabase(final String databaseName) throws CantOpenDatabaseException ,
-                                                                         DatabaseNotFoundException {
-
-        try {
-            AndroidDatabase database;
-            String hasDBName = hashDataBaseName(databaseName);
-            database = new AndroidDatabase(context.getFilesDir().getPath(), hasDBName);
-            database.openDatabase();
-            return database;
-        } catch (final NoSuchAlgorithmException e){
-
-            throw new CantOpenDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (final Exception e){
-
-            throw new CantOpenDatabaseException(e, null, "Unhandled Exception.");
-        }
-    }
-
-    @Override
-    public final Database createDatabase (final String databaseName) throws CantCreateDatabaseException {
-
-        try {
-            AndroidDatabase database;
-            String hasDBName = hashDataBaseName(databaseName);
-            database = new AndroidDatabase(context.getFilesDir().getPath(), hasDBName);
-            database.createDatabase(hasDBName);
-            return database;
-        }
-        catch (final NoSuchAlgorithmException e){
-
-            throw new CantCreateDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (Exception e){
-
-            throw new CantCreateDatabaseException(e, null, "Unhandled Exception.");
-        }
-
-    }
-
-    @Override
-    public final void deleteDatabase(final String databaseName) throws CantOpenDatabaseException ,
-                                                                       DatabaseNotFoundException {
-        try{
-            String hasDBName = hashDataBaseName(databaseName);
-            AndroidDatabase database;
-            database = new AndroidDatabase(context.getFilesDir().getPath(), hasDBName);
-            database.deleteDatabase();
-
-        } catch (final NoSuchAlgorithmException e){
-
-            throw new CantOpenDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (final Exception e){
-
-            throw new CantOpenDatabaseException(e, null, "Unhandled Exception.");
-        }
-    }
-/*
-    private String hashDataBaseName(final String databaseName) throws CantHashDatabaseNameException {
+    public final void start() throws CantStartPluginException {
 
         try {
 
-            final MessageDigest md = MessageDigest.getInstance("SHA-256");
-            md.update(databaseName.getBytes(Charset.forName("UTF-8")));
-            byte[] digest = md.digest();
-            byte[] encoded = Base64.encode(digest, 1);
+            platformDatabaseSystemManager = new AndroidPlatformDatabaseSystem(context);
 
-            String encryptedString = new String(encoded, "UTF-8");
+            super.start();
 
-            encryptedString = encryptedString.replace("/","");
+        } catch (final Exception e) {
 
-            return encryptedString.replace("\n","");
-
-        } catch(final UnsupportedEncodingException e){
-
-            throw new CantHashDatabaseNameException(e, "databaseName: "+ databaseName, "Unsupported encoding exception.");
-        } catch(final NoSuchAlgorithmException e){
-
-            throw new CantHashDatabaseNameException(e, "databaseName: "+ databaseName, "No such algorithm.");
+            throw new CantStartPluginException(e, "Platform Database System Manager starting.", "Unhandled Exception trying to start the Platform Database System manager.");
         }
-    }*/
-    private String hashDataBaseName(String databaseName) throws NoSuchAlgorithmException {
-        String encryptedString = databaseName;
-        try{
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            md.update(databaseName.getBytes(Charset.forName("UTF-8")));
-            byte[] digest = md.digest();
-            byte[] encoded = Base64.encode(digest, 1);
-
-            try {
-                encryptedString = new String(encoded, "UTF-8");
-            } catch (Exception e) {
-                throw new NoSuchAlgorithmException (e);
-            }
-
-
-        }catch(NoSuchAlgorithmException e){
-            throw e;
-        }
-        encryptedString = encryptedString.replace("+","");
-        return encryptedString.replace("/","");
     }
+
+    @Override
+    public final FermatManager getManager() {
+        return platformDatabaseSystemManager;
+    }
+
 }
