+++ OSA/addon/android/fermat-osa-addon-android-database-system-bitdubai/src/main/java/com/bitdubai/fermat_osa_addon/layer/android/database_system/developer/bitdubai/version_1/PluginDatabaseSystemAndroidAddonRoot.java
@@ -1,24 +1,14 @@
 package com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1;
 
 import android.content.Context;
-import android.util.Base64;
 
 import com.bitdubai.fermat_api.CantStartPluginException;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.abstract_classes.AbstractAddon;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededOsContext;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.AddonVersionReference;
-import com.bitdubai.fermat_api.layer.all_definition.enums.ServiceStatus;
 import com.bitdubai.fermat_api.layer.all_definition.util.Version;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.Database;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.PluginDatabaseSystem;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantCreateDatabaseException;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantOpenDatabaseException;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.DatabaseNotFoundException;
-import com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1.structure.AndroidDatabase;
-
-import java.nio.charset.Charset;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.UUID;
+import com.bitdubai.fermat_osa_addon.layer.android.database_system.developer.bitdubai.version_1.structure.AndroidPluginDatabaseSystem;
 
 /**
  * This addon handles a layer of database representation.
@@ -29,122 +19,37 @@ import java.util.UUID;
  * Modified by lnacosta (laion.cj91@gmail.com) on 26/10/2015.
  */
 
-public final class PluginDatabaseSystemAndroidAddonRoot extends AbstractAddon implements PluginDatabaseSystem {
+public final class PluginDatabaseSystemAndroidAddonRoot extends AbstractAddon {
 
+    @NeededOsContext
     private Context context;
 
+    private FermatManager pluginDatabaseSystemManager;
 
+    /**
+     * Constructor without parameters.
+     */
     public PluginDatabaseSystemAndroidAddonRoot() {
-        super(
-                new AddonVersionReference(new Version()),
-                true // true because uses system context (the core will assign the context if needed)
-        );
+        super(new AddonVersionReference(new Version()));
     }
 
     @Override
     public final void start() throws CantStartPluginException {
 
-        if (this.getOsContext() != null && this.getOsContext() instanceof Context) {
-            context = (Context) this.getOsContext();
-            this.serviceStatus = ServiceStatus.STARTED;
-        } else {
-            throw new CantStartPluginException(
-                    "osContext: "+this.getOsContext(),
-                    "Context is not instance of Android Context or is null."
-            );
-        }
-    }
-    /**
-     * PluginDatabaseSystem Interface implementation.
-     */
-
-    @Override
-    public final Database openDatabase(final UUID   ownerId     ,
-                                       final String databaseName) throws CantOpenDatabaseException ,
-                                                                   DatabaseNotFoundException {
-
         try {
 
-            String hasDBName = hashDataBaseName(databaseName);
-            AndroidDatabase database = new AndroidDatabase(context.getFilesDir().getPath(), ownerId, hasDBName);
-            database.openDatabase();
-            return database;
+            pluginDatabaseSystemManager = new AndroidPluginDatabaseSystem(context);
 
-        } catch (final NoSuchAlgorithmException e) {
-
-            throw new CantOpenDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (final DatabaseNotFoundException exception){
-
-            throw exception;
-        } catch (final Exception e){
-
-            throw new CantOpenDatabaseException(e, null, "Unhandled Exception");
-        }
-
-    }
-
-    @Override
-    public final void deleteDatabase(final UUID   ownerId     ,
-                                     final String databaseName) throws CantOpenDatabaseException ,
-                                                                 DatabaseNotFoundException {
-
-        try {
-
-            String hasDBName = hashDataBaseName(databaseName);
-            AndroidDatabase database = new AndroidDatabase(context.getFilesDir().getPath(), ownerId, hasDBName);
-            database.deleteDatabase();
-
-        } catch (final NoSuchAlgorithmException e){
-
-            throw new CantOpenDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (final DatabaseNotFoundException exception){
-
-            throw exception;
-        } catch (final Exception e){
-
-            throw new CantOpenDatabaseException(e, null, "Unhandled Exception");
-        }
-
-    }
-
-    @Override
-    public final Database createDatabase(final UUID   ownerId     ,
-                                         final String databaseName) throws CantCreateDatabaseException {
-
-        try {
-            String hasDBName = hashDataBaseName(databaseName);
-            AndroidDatabase database = new AndroidDatabase(context.getFilesDir().getPath(), ownerId, hasDBName);
-            database.createDatabase(hasDBName);
-            return database;
-
-        } catch (final NoSuchAlgorithmException e){
-
-            throw new CantCreateDatabaseException(e, "Database Name : " + databaseName, "This is a hash failure, we have to check the hashing algorithm used for the generation of the Hashed Database Name");
-        } catch (final Exception e){
-
-            throw new CantCreateDatabaseException(e, null, "Unhandled Exception");
-        }
-
-    }
-
-    private String hashDataBaseName(final String databaseName) throws NoSuchAlgorithmException {
-
-        MessageDigest md = MessageDigest.getInstance("SHA-256");
-        md.update(databaseName.getBytes(Charset.forName("UTF-8")));
-        byte[] digest = md.digest();
-        byte[] encoded = Base64.encode(digest, 1);
-
-        try {
-
-            String encryptedString = new String(encoded, "UTF-8");
-
-            encryptedString = encryptedString.replace("/","");
-
-            return encryptedString.replace("\n","");
+            super.start();
 
         } catch (final Exception e) {
 
-            throw new NoSuchAlgorithmException (e);
+            throw new CantStartPluginException(e, "Plugin Database System Manager starting.", "Unhandled Exception trying to start the Plugin Database System manager.");
         }
+    }
+
+    @Override
+    public final FermatManager getManager() {
+        return pluginDatabaseSystemManager;
     }
 }
