+++ OSA/addon/android/fermat-osa-addon-android-file-system-bitdubai/src/main/java/com/bitdubai/fermat_osa_addon/layer/android/file_system/developer/bitdubai/version_1/structure/AndroidPlatformFileSystem.java
@@ -3,7 +3,6 @@ package com.bitdubai.fermat_osa_addon.layer.android.file_system.developer.bitdub
 import android.content.Context;
 import android.util.Base64;
 
-import com.bitdubai.fermat_api.FermatException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.FileLifeSpan;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.FilePrivacy;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.PlatformBinaryFile;
@@ -12,8 +11,8 @@ import com.bitdubai.fermat_api.layer.osa_android.file_system.PlatformTextFile;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantCreateFileException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantLoadFileException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.FileNotFoundException;
+import com.bitdubai.fermat_osa_addon.layer.android.file_system.developer.bitdubai.version_1.exceptions.CantHashFileNameException;
 
-import java.io.Serializable;
 import java.nio.charset.Charset;
 import java.security.MessageDigest;
 
@@ -26,7 +25,7 @@ import java.security.MessageDigest;
  * The Platform File System is the implementation of the file system that is handled to external plugins not requires the plug in to identify itself.
  */
 
-public class AndroidPlatformFileSystem implements PlatformFileSystem,Serializable {
+public class AndroidPlatformFileSystem implements PlatformFileSystem  {
 
     private static final String CHARSET_NAME = "UTF-8";
     private static final String DIGEST_ALGORITHM = "SHA-256";
@@ -35,113 +34,130 @@ public class AndroidPlatformFileSystem implements PlatformFileSystem,Serializabl
      * PlatformFileSystem interface member variables.
      */
 
-    private String contextPath;
+    private final Context context;
 
-    public AndroidPlatformFileSystem(String contextPath) {
-        this.contextPath = contextPath;
+    public AndroidPlatformFileSystem(final Context context) {
+        this.context = context;
     }
 
-    /**
-     * PlatformFileSystem interface implementation.
-     */
 
-    /**
-     *<p>This method gets a new PlatformTextFile object. And load file content on memory
-     *
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to load
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PlatformTextFile object
-     * @throws FileNotFoundException
-     * @throws CantCreateFileException
-     */
     @Override
-    public PlatformTextFile getFile(String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws FileNotFoundException,CantCreateFileException {
-        checkContext();
+    public final PlatformTextFile getFile(final String       directoryName,
+                                          final String       fileName     ,
+                                          final FilePrivacy  privacyLevel ,
+                                          final FileLifeSpan lifeSpan     ) throws FileNotFoundException   ,
+            CantCreateFileException {
+
         try {
-            AndroidPlatformTextFile newFile = new AndroidPlatformTextFile( contextPath, directoryName,hashFileName(fileName), privacyLevel, lifeSpan);
+
+            final AndroidPlatformTextFile newFile = new AndroidPlatformTextFile(
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
             newFile.loadFromMedia();
             return newFile;
         }
         catch (CantLoadFileException exception){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE,exception,null,"Check the cause of this error");
-        }
-        catch(Exception exception){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE,FermatException.wrapException(exception),null,"Check the cause of this error");
+
+            throw new FileNotFoundException(exception, null, "Check the cause of this error");
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e,"", "Unhandled error.");
         }
 
     }
 
-    /**
-     *<p>This method create a new PlatformTextFile object.
-     *
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to load
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PlatformTextFile object
-     * @throws CantCreateFileException
-     */
     @Override
-    public PlatformTextFile createFile(String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws  CantCreateFileException{
-        checkContext();
-       try{
-        return new AndroidPlatformTextFile( contextPath,directoryName,hashFileName(fileName), privacyLevel, lifeSpan);
-       }catch (CantCreateFileException exception){
-               throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE,exception,null,"Check the cause of this error");
-           }
-           catch(Exception e){
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
-        }
-    }
+    public final PlatformTextFile createFile(final String       directoryName,
+                                             final String       fileName     ,
+                                             final FilePrivacy  privacyLevel ,
+                                             final FileLifeSpan lifeSpan     ) throws  CantCreateFileException {
 
-    @Override
-    public PlatformBinaryFile getBinaryFile(String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws FileNotFoundException,CantCreateFileException {
-        checkContext();
         try {
-            AndroidPlatformBinaryFile newFile = new AndroidPlatformBinaryFile(contextPath, directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
+
+            return new AndroidPlatformTextFile(
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e,"", "Unhandled error.");
+        }
+    }
+
+    @Override
+    public final PlatformBinaryFile getBinaryFile(final String       directoryName,
+                                                  final String       fileName     ,
+                                                  final FilePrivacy  privacyLevel ,
+                                                  final FileLifeSpan lifeSpan     ) throws FileNotFoundException   ,
+            CantCreateFileException {
+
+        try {
+
+            final AndroidPlatformBinaryFile newFile = new AndroidPlatformBinaryFile(
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
             newFile.loadFromMedia();
             return newFile;
+
         } catch (CantLoadFileException e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, e, "", "Check the cause");
+
+            throw new FileNotFoundException(e, "", "Check the cause");
         }catch(Exception e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
+
+            throw new CantCreateFileException(e, "", "Unhandled error.");
         }
     }
 
     @Override
-    public PlatformBinaryFile createBinaryFile(String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws CantCreateFileException {
-       checkContext();
+    public final PlatformBinaryFile createBinaryFile(final String       directoryName,
+                                                     final String       fileName     ,
+                                                     final FilePrivacy  privacyLevel ,
+                                                     final FileLifeSpan lifeSpan     ) throws CantCreateFileException {
+
         try {
-        return new AndroidPlatformBinaryFile(contextPath, directoryName,hashFileName(fileName), privacyLevel, lifeSpan);
-        }catch (CantCreateFileException exception){
-            throw exception;
-        }catch(Exception e){
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
-           }
+
+            return new AndroidPlatformBinaryFile(
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e, "", "Unhandled error.");
+        }
     }
 
-       /**
-     *
-     * Hash the file name using the algorithm SHA 256
-     */
+    private String hashFileName(final String fileName) throws CantHashFileNameException {
 
-    private String hashFileName(String fileName) throws CantCreateFileException {
-        try{
-            MessageDigest md = MessageDigest.getInstance(DIGEST_ALGORITHM);
+        try {
+
+            final MessageDigest md = MessageDigest.getInstance(DIGEST_ALGORITHM);
             md.update(fileName.getBytes(Charset.forName(CHARSET_NAME)));
             byte[] digest = md.digest();
             byte[] encoded = Base64.encode(digest,1);
-            String encryptedString = new String(encoded, CHARSET_NAME);
+            final String encryptedString = new String(encoded, CHARSET_NAME);
             return encryptedString.replace("/","");
-        }catch(Exception e){
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "This Should never happen unless we change the DIGEST_ALGORITHM Constant");
-        }
-    }
 
-    private void checkContext() throws CantCreateFileException {
-        if(contextPath == null)
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, null, "Context: null", "Context can't ne Null, you need to call setContext before using the FileSystem");
+        } catch(Exception e){
+
+            throw new CantHashFileNameException(e, "", "This Should never happen unless we change the DIGEST_ALGORITHM Constant");
+        }
     }
 }
