+++ OSA/addon/android/fermat-osa-addon-android-file-system-bitdubai/src/main/java/com/bitdubai/fermat_osa_addon/layer/android/file_system/developer/bitdubai/version_1/structure/AndroidPluginFileSystem.java
@@ -3,7 +3,6 @@ package com.bitdubai.fermat_osa_addon.layer.android.file_system.developer.bitdub
 import android.content.Context;
 import android.util.Base64;
 
-import com.bitdubai.fermat_api.FermatException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.FileLifeSpan;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.FilePrivacy;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginBinaryFile;
@@ -12,6 +11,7 @@ import com.bitdubai.fermat_api.layer.osa_android.file_system.PluginTextFile;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantCreateFileException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantLoadFileException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.FileNotFoundException;
+import com.bitdubai.fermat_osa_addon.layer.android.file_system.developer.bitdubai.version_1.exceptions.CantHashFileNameException;
 
 import java.nio.charset.Charset;
 import java.security.MessageDigest;
@@ -36,165 +36,192 @@ public class AndroidPluginFileSystem implements PluginFileSystem {
      * PluginFileSystem interface member variables.
      */
     
-    private String contextPath;
+    private Context context;
 
-    public AndroidPluginFileSystem(String contextPath) {
-        this.contextPath = contextPath;
+    public AndroidPluginFileSystem(Context context) {
+        this.context = context;
     }
-/**
-     * PluginFileSystem interface implementation.
-     */
 
-    /**
-     *<p>This method return a PluginTextFile object and load file content on memory
-     *
-     * @param ownerId  PlugIn id
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to load
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PluginTextFile object
-     * @throws FileNotFoundException
-     * @throws CantCreateFileException
-     */
     @Override
-    public PluginTextFile getTextFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws FileNotFoundException,CantCreateFileException {
-        checkContext();
+    public final PluginTextFile getTextFile(final UUID         ownerId      ,
+                                            final String       directoryName,
+                                            final String       fileName     ,
+                                            final FilePrivacy  privacyLevel ,
+                                            final FileLifeSpan lifeSpan     ) throws FileNotFoundException   ,
+            CantCreateFileException {
+
         try {
-            //execute AndroidPluginTextFile constructor
-            AndroidPluginTextFile newFile = new AndroidPluginTextFile(ownerId, contextPath,directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
-            //load content file
+            AndroidPluginTextFile newFile = new AndroidPluginTextFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
             newFile.loadFromMedia();
             return newFile;
         }
-        catch (CantLoadFileException e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, e, "", "Check the cause");
+        catch (CantLoadFileException exception){
+
+            throw new FileNotFoundException(exception, null, "Check the cause of this error");
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e,"", "Unhandled error.");
         }
-        catch (Exception e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "Check the cause");
-        }
+
     }
 
-    /**
-     *<p>This method create a new PluginTextFile object.
-     *
-     * @param ownerId PlugIn id
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to manage
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PluginTextFile object
-     * @throws CantCreateFileException
-     */
     @Override
-    public PluginTextFile createTextFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws CantCreateFileException{
-        checkContext();
-     try{
-        return new AndroidPluginTextFile(ownerId, contextPath, directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
-       }catch(CantCreateFileException e){
-            throw e;
-            // throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
-        }catch(Exception e){
-         throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
-     }
-   
-    }
+    public final PluginTextFile createTextFile(final UUID         ownerId      ,
+                                               final String       directoryName,
+                                               final String       fileName     ,
+                                               final FilePrivacy  privacyLevel ,
+                                               final FileLifeSpan lifeSpan     ) throws  CantCreateFileException {
 
-    /**
-     *<p>This method return a PluginBinaryFile object and load file content on memory
-     *
-     * @param ownerId PlugIn id
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to load
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PluginBinaryFile object
-     * @throws FileNotFoundException
-     * @throws CantCreateFileException
-     */
-    @Override
-    public PluginBinaryFile getBinaryFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws FileNotFoundException,CantCreateFileException{
-        checkContext();
         try {
-            AndroidPluginBinaryFile newFile = new AndroidPluginBinaryFile(ownerId, contextPath, directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
+
+            return new AndroidPluginTextFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e,"", "Unhandled error.");
+        }
+    }
+
+    @Override
+    public final PluginBinaryFile getBinaryFile(final UUID         ownerId      ,
+                                                final String       directoryName,
+                                                final String       fileName     ,
+                                                final FilePrivacy  privacyLevel ,
+                                                final FileLifeSpan lifeSpan     ) throws FileNotFoundException   ,
+            CantCreateFileException {
+
+        try {
+
+            final AndroidPluginBinaryFile newFile = new AndroidPluginBinaryFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
             newFile.loadFromMedia();
             return newFile;
+
         } catch (CantLoadFileException e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, e, "", "Check the cause");
+
+            throw new FileNotFoundException(e, "", "Check the cause");
+        }catch(Exception e){
+
+            throw new CantCreateFileException(e, "", "Unhandled error.");
+        }
+    }
+
+    @Override
+    public final PluginBinaryFile createBinaryFile(final UUID         ownerId      ,
+                                                   final String       directoryName,
+                                                   final String       fileName     ,
+                                                   final FilePrivacy  privacyLevel ,
+                                                   final FileLifeSpan lifeSpan     ) throws CantCreateFileException {
+
+        try {
+
+            return new AndroidPluginBinaryFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+
+        } catch(Exception e){
+
+            throw new CantCreateFileException(e, "", "Unhandled error.");
+        }
+    }
+
+    @Override
+    public final void deleteTextFile(final UUID         ownerId      ,
+                                     final String       directoryName,
+                                     final String       fileName     ,
+                                     final FilePrivacy  privacyLevel ,
+                                     final FileLifeSpan lifeSpan     ) throws CantCreateFileException,
+            FileNotFoundException  {
+
+        try {
+
+            final AndroidPluginTextFile newFile =  new AndroidPluginTextFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
+            newFile.delete();
+
         } catch (Exception e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "Check the cause");
+
+            throw new CantCreateFileException(e, "", "Check the cause");
         }
-    }
 
-    /**
-     *<p>This method create a new PluginBinaryFile object.
-     *
-     * @param ownerId PlugIn id
-     * @param directoryName name of the directory where the files are stored
-     * @param fileName name of file to load
-     * @param privacyLevel level of privacy for the file, if it is public or private
-     * @param lifeSpan lifeSpan of the file, whether it is permanent or temporary
-     * @return PluginBinaryFile object
-     * @throws CantCreateFileException
-     */
-
-    @Override
-    public PluginBinaryFile createBinaryFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws CantCreateFileException{
-        checkContext();
-       try{
-        return new AndroidPluginBinaryFile(ownerId, contextPath, directoryName,hashFileName(fileName), privacyLevel, lifeSpan);
-          }catch(Exception e){
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e),"", "Check the cause of this error");
-         }
     }
 
     @Override
-    public void deleteTextFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws CantCreateFileException, FileNotFoundException {
+    public final void deleteBinaryFile(final UUID         ownerId      ,
+                                       final String       directoryName,
+                                       final String       fileName     ,
+                                       final FilePrivacy  privacyLevel ,
+                                       final FileLifeSpan lifeSpan     ) throws CantCreateFileException,
+                                                                                FileNotFoundException  {
+
         try {
-            //execute AndroidPluginTextFile constructor
-            AndroidPluginTextFile newFile = new AndroidPluginTextFile(ownerId, contextPath,directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
+
+            final AndroidPluginBinaryFile newFile = new AndroidPluginBinaryFile(
+                    ownerId                        ,
+                    context.getFilesDir().getPath(),
+                    directoryName                  ,
+                    hashFileName(fileName)         ,
+                    privacyLevel                   ,
+                    lifeSpan
+            );
             newFile.delete();
-        }
-        catch (Exception e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "Check the cause");
-        }
 
-    }
+        } catch (FileNotFoundException e){
 
-    @Override
-    public void deleteBinaryFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws CantCreateFileException, FileNotFoundException{
-        try {
-            //execute AndroidPluginTextFile constructor
-            AndroidPluginBinaryFile newFile = new AndroidPluginBinaryFile(ownerId, contextPath,directoryName, hashFileName(fileName), privacyLevel, lifeSpan);
-            newFile.delete();
-        }
-        catch (Exception e){
-            throw new FileNotFoundException(FileNotFoundException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "Check the cause");
+            throw e;
+        } catch (Exception e){
+
+            throw new CantCreateFileException(e, "", "Check the cause");
         }
     }
 
-    /**
-     *
-     * Hash the file name using the algorithm SHA 256
-     */
+    private String hashFileName(final String fileName) throws CantHashFileNameException {
 
-    private String hashFileName(String fileName) throws CantCreateFileException{
         try {
-            MessageDigest md = MessageDigest.getInstance(DIGEST_ALGORITHM);
+
+            final MessageDigest md = MessageDigest.getInstance(DIGEST_ALGORITHM);
             md.update(fileName.getBytes(Charset.forName(CHARSET_NAME)));
             byte[] digest = md.digest();
-            byte[] encoded = Base64.encode(digest, 1);
-            String encryptedString = new String(encoded, CHARSET_NAME);
-            encryptedString = encryptedString.replace("/","");
-            return encryptedString.replace("\n","");
-        } catch (Exception e) {
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, FermatException.wrapException(e), "", "This Should never happen unless we change the DIGEST_ALGORITHM Constant");
+            byte[] encoded = Base64.encode(digest,1);
+            final String encryptedString = new String(encoded, CHARSET_NAME);
+            return encryptedString.replace("/","");
+
+        } catch(Exception e){
+
+            throw new CantHashFileNameException(e, "", "This Should never happen unless we change the DIGEST_ALGORITHM Constant");
         }
-
-    }
-
-    private void checkContext() throws CantCreateFileException {
-        if(contextPath == null)
-            throw new CantCreateFileException(CantCreateFileException.DEFAULT_MESSAGE, null, "Context: null", "Context can't ne Null, you need to call setContext before using the FileSystem");
     }
 }
