+++ PIP/library/api/fermat-pip-api/src/main/java/com/bitdubai/fermat_pip_api/layer/platform_service/error_manager/ErrorManager.java
@@ -1,5 +1,6 @@
 package com.bitdubai.fermat_pip_api.layer.platform_service.error_manager;
 
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.AddonVersionReference;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginVersionReference;
 import com.bitdubai.fermat_api.layer.all_definition.enums.Addons;
@@ -14,7 +15,7 @@ import com.bitdubai.fermat_api.layer.dmp_engine.sub_app_runtime.enums.SubApps;
  * Created by ciencias on 05.02.15.
  * Modified by lnacosta (laion.cj91@gmail.com) on 26/10/2015.
  */
-public interface ErrorManager {
+public interface ErrorManager extends FermatManager {
 
     void reportUnexpectedPlatformException(PlatformComponents exceptionSource, UnexpectedPlatformExceptionSeverity unexpectedPlatformExceptionSeverity, Exception exception);
 
@@ -47,7 +48,7 @@ public interface ErrorManager {
      * @param exception          exception reported.
      */
     void reportUnexpectedAddonsException(final AddonVersionReference             exceptionSource  ,
-                                         final UnexpectedPluginExceptionSeverity exceptionSeverity,
+                                         final UnexpectedAddonsExceptionSeverity exceptionSeverity,
                                          final Exception                         exception        );
 
     /**
