+++ PIP/library/api/fermat-pip-api/src/main/java/com/bitdubai/fermat_pip_api/layer/platform_service/platform_info/interfaces/PlatformInfoManager.java
@@ -1,10 +1,8 @@
 package com.bitdubai.fermat_pip_api.layer.platform_service.platform_info.interfaces;
 
-import com.bitdubai.fermat_api.layer.all_definition.enums.Platforms;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_pip_api.layer.platform_service.platform_info.interfaces.exceptions.CantLoadPlatformInformationException;
 import com.bitdubai.fermat_pip_api.layer.platform_service.platform_info.interfaces.exceptions.CantSetPlatformInformationException;
-
-import java.util.List;
 
 /**
  * The Interface <code>PlatformInfoManager</code>
@@ -16,7 +14,7 @@ import java.util.List;
  * @since Java JDK 1.7
  */
 
-public interface PlatformInfoManager {
+public interface PlatformInfoManager extends FermatManager {
 
     PlatformInfo getPlatformInfo() throws CantLoadPlatformInformationException;
