+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/osa_android/database_system/PluginDatabaseSystem.java
@@ -1,8 +1,9 @@
 package com.bitdubai.fermat_api.layer.osa_android.database_system;
 
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.DatabaseNotFoundException;
-import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantOpenDatabaseException;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantCreateDatabaseException;
+import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.CantOpenDatabaseException;
+import com.bitdubai.fermat_api.layer.osa_android.database_system.exceptions.DatabaseNotFoundException;
 
 import java.util.UUID;
 
@@ -18,7 +19,7 @@ import java.util.UUID;
  *  @since   18/01/15.
  * */
 
- public interface PluginDatabaseSystem {
+ public interface PluginDatabaseSystem extends FermatManager {
 
     Database openDatabase(UUID ownerId, String databaseName) throws CantOpenDatabaseException, DatabaseNotFoundException;
