+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/osa_android/file_system/PluginFileSystem.java
@@ -1,5 +1,6 @@
 package com.bitdubai.fermat_api.layer.osa_android.file_system;
 
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.CantCreateFileException;
 import com.bitdubai.fermat_api.layer.osa_android.file_system.exceptions.FileNotFoundException;
 
@@ -17,7 +18,7 @@ import java.util.UUID;
  *  @since   22/01/15.
  * */
 
- public interface PluginFileSystem {
+ public interface PluginFileSystem extends FermatManager {
 
     public PluginTextFile getTextFile(UUID ownerId, String directoryName, String fileName, FilePrivacy privacyLevel, FileLifeSpan lifeSpan) throws FileNotFoundException,CantCreateFileException;
